Move helper.externals to helpers
Fix rebase
